+++ NEWS.md
@@ -1,5 +1,6 @@
 # ctrdata 1.25.1.9000
 
+- For `dbFindIdsUniqueTrials()` with CTIS, only include trial record with highest resubmission number
 - Handle expired CDN deep links
 
 # ctrdata 1.25.1